Added Stash system
you should now be able to save stuff inside of your stash (and not be able to take stuff out of other people's stashes)

+++ gamemode/sv_stash.lua
@@ -1,28 +1,123 @@
 
-function GM:Initialize()
+util.AddNetworkString( "RequestStash" )
+util.AddNetworkString( "SendStash" )
 
-    -- table contains variables IsStackable(bool), ItemName(string), ItemCount(int), ItemOwner(Steam64)
+util.AddNetworkString( "TakeFromStash" )
+util.AddNetworkString( "PutWepInStash" )
 
-    sql.Query( "CREATE TABLE IF NOT EXISTS stash_table ( IsStackable INTEGER, ItemName TEXT, ItemCount INTEGER, ItemOwner INTEGER )" )
+util.AddNetworkString( "StashMenuReload" )
+
+local function SendStashToClient(player)
+
+    local value = sql.Query( "SELECT * FROM stash_table WHERE ItemOwner = " .. player:SteamID64() .. ";" )
+
+    print("stash request received")
+
+    if value != nil then
+
+        PrintTable(value)
+
+        net.Start( "SendStash" )
+        net.WriteTable( value )
+        net.Send(player)
+
+    end
 
 end
 
+net.Receive("RequestStash",function (len, ply)
 
-local function ReadStash()
+    print("got request to show stash")
+    SendStashToClient(ply)
+	
+end)
 
-end
+net.Receive("PutWepInStash",function (len, ply)
+
+    local item = net.ReadString()
+    local count = 1
+    local type = SQLStr("wep", false)
+
+    if ply:HasWeapon( item ) == false then print("Player does not have " .. item .. "!") return end
+
+    local highestID = sql.QueryValue( "SELECT MAX(UniqueID) FROM stash_table;" )
+
+    if highestID == "NULL" then highestID = 1 end
+
+    print("The highest ID found in stash_table is " .. highestID)
+
+    sql.Query( "INSERT INTO stash_table ( ItemName, ItemCount, ItemType, ItemOwner, UniqueID ) VALUES( " .. SQLStr(item, false) .. ", " .. count .. ", " .. type .. ", " .. ply:SteamID64() .. ", " .. highestID + 1 ..")" )
+    print("Added into SQL Table (stash_table): " .. SQLStr(item, false) .. " " .. count .. " " .. type .. " " .. ply:SteamID64() .. " ".. highestID + 1 .."!")
+    print("Last SQL Error = " .. tostring(sql.LastError()))
+
+    ply:StripWeapon( item )
+
+    net.Start("StashMenuReload")
+    net.Send(ply)
+	
+end)
+
+net.Receive("TakeFromStash",function (len, ply)
+
+    requestedItemName = net.ReadString()
+    stashItemName = sql.QueryValue( "SELECT ItemName FROM stash_table WHERE ItemName = " .. sql.SQLStr(requestedItemName) .. ";" )
+    
+    if(stashItemName != nil) then
+
+        local lowestID = sql.QueryValue( "SELECT UniqueID FROM stash_table WHERE ItemName = " .. sql.SQLStr(stashItemName) .. " AND ItemOwner = " .. ply:SteamID64() .. " AND UniqueID = ( SELECT MIN(UniqueID) FROM stash_table WHERE ItemName = " .. sql.SQLStr(stashItemName) .. ");" )
+        print("The lowest ID found in stash_table is" .. lowestID)
+
+        sql.Query( "DELETE FROM stash_table WHERE UniqueID = " .. lowestID .. ";" )
+
+        ply:Give(stashItemName, true)
+
+        net.Start( "StashMenuReload" )
+        net.Send(ply)
+    end
+
+end)
 
 
-local function WriteStashIndex()
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+-- Debug Console Commands
+
+local function CreateTable(ply)
+
+    sql.Query( "CREATE TABLE IF NOT EXISTS stash_table ( ItemName TEXT, ItemCount INTEGER, ItemType TEXT, ItemOwner INTEGER, UniqueID INTEGER )" )
+    
 end
 
 
 local function ConsoleReturnStashContents(ply, cmd, args)
 
-    local value = sql.Query( "SELECT * FROM stash_table ;" )
-    PrintTable(value)
-    print("Last SQL Error = " .. tostring(sql.LastError()))
+    if ply:IsAdmin() == false then return end
+
+    local value = sql.Query( "SELECT * FROM stash_table;" )
+
+    if(value != false or nil) then
+        PrintTable(value)
+        print("Last SQL Error = " .. tostring(sql.LastError()))
+    end
 
 end
 concommand.Add("read_stash", ConsoleReturnStashContents)
@@ -30,15 +125,19 @@ concommand.Add("read_stash", ConsoleReturnStashContents)
 
 local function ConsoleWriteStashContents(ply, cmd, args)
 
-    local stackable = tostring(tobool(args[1]))
-    local item = SQLStr(args[2], false)
-    local count = args[3]
-    local owner = ply:SteamID64()
+    if ply:IsAdmin() == false then return end
 
-    print(stackable .. "\n" .. item .. "\n" .. count .. "\n" .. owner)
+    local item = SQLStr(args[1], false)
+    local count = args[2]
+    local type = SQLStr(args[3], false)
 
-    local returnedTable = sql.Query( "INSERT INTO stash_table ( IsStackable, ItemName, ItemCount, ItemOwner ) VALUES( " .. stackable .. ", " .. item .. ", " .. count .. ", " .. owner ..")" )
-    print(returnedTable)
+    local highestID = sql.QueryValue( "SELECT MAX(UniqueID) FROM stash_table;" )
+
+    if highestID == "NULL" then highestID = 1 end
+
+    print("The highest ID found in stash_table is" .. highestID)
+
+    sql.Query( "INSERT INTO stash_table ( ItemName, ItemCount, ItemType, ItemOwner, UniqueID ) VALUES( " .. item .. ", " .. count .. ", " .. type .. ", " .. ply:SteamID64() .. ", " .. highestID + 1 ..")" )
     print("Last SQL Error = " .. tostring(sql.LastError()))
 
 end
@@ -46,10 +145,22 @@ concommand.Add("write_stash", ConsoleWriteStashContents)
 
 local function WipeTable(ply, cmd, args)
 
-    local returnedTable = sql.Query( " DELETE FROM stash_table; " )
+    if ply:IsAdmin() == false then return end
 
-    print(returnedTable)
+    sql.Query( " DELETE FROM stash_table; " )
+
     print("Last SQL Error = " .. tostring(sql.LastError()))
 
 end
 concommand.Add("delete_stash", WipeTable)
+
+local function ConsoleDropTable(ply, cmd, args)
+
+    if ply:IsAdmin() == false then return end
+
+    sql.Query( "DROP TABLE stash_table;" )
+
+    CreateTable(ply)
+
+end
+concommand.Add("reset_stash", ConsoleDropTable)
